macosx: audio filters: load audio-filters variable from preset correctly
Load audio filters correctly. Previously, only new filters were
added, but removed filters not disabled.
playlist_EnableAudioFilter ensures correct filter order in addition
to modifying the audio-filters variable. As the profile string is
created and changed with playlist_EnableAudioFilter, the saved
string conforms to the rules, so here it can be directly loaded.

Also removes redundant UI updates, which are done in a different
method already.

refs #19260