contrib: flac: Don't force __MSVCRT_VERSION__ to a specific (old) version
The forcing of this define was added upstream many years ago, probably
to expose certain features that aren't available by default in the old
mingw.org headers. In mingw-w64, setting __MSVCRT_VERSION__ doesn't
expose much new functionality at all (contrary to in mingw.org), and
the default in mingw-w64 is 0x0700.

Therefore, since we're using mingw-w64 anyway, don't force this version
but just use whatever the toolchain default is, assuming the default
is new enough.

This fixes building with mingw-w64 with the default msvcrt set to
ucrtbase.