direct3d11: switch the pixel shader to point interpolation when reaching visible sides
The texture is bigger than the visible sides, so if we use linear interpolation
on the last pixel it will use values with unknown values on the part where it
would do clamping otherwise. So we force it to use point interpolation (not
using pixels around) for the last line/column to display.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit bf2eb2ef893c23dabbd7c7e835d417d83ae4b054)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>