contribs: qt: Clean up libQt5Bootstrap.a in the build tree as well
When building the Qt tools for the target architecture in cross
compilation, the bootstrap object library is rebuilt. (In cross
builds of Qt, the tools are only built for the build host.
The VLC contribs build them manually for the targeted environment
afterwards.) The tools bootstrap library wasn't removed inbetween
though.

This meant that the object files for the second round (for the
cross target environment) were added to the bootstrap library for
the host environment. Most object files were just replaced, but some
few object files weren't built at all in cross builds for windows.
This meant that after cross-building the bootstrap library for windows,
it still contained a few object files for the native host build
environment as well.

lld-link errors out when it encounters object files in a static library
that it can't parse (in this case, ELF files when doing a windows cross
build).

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>