contribs: Make sure either USE_FFMPEG or USE_LIBAV is set
The rest of rules.mak assumes that either of them is set (and check
either "ifdef USE_FFMPEG" or "ifdef USE_LIBAV"), but by default
neither is set. If using the "ifndef USE_LIBAV" branch and fetching
sources from ffmpeg, make sure USE_FFMPEG is set in the rest of the
makefile.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>