opengl: add support for dithering
This uses libplacebo's high quality blue noise algorithm. The lut_size
is chosen as 4 because the method of embedding (literal LUT) is not that
great. Larger texture sizes would require OpenGL/RA support, which is
not currently in the scope of VLC+libplacebo. (Alternatively, we could
try and use a uniform float array instead of a constant array - but it
doesn't seem to matter much)

In theory, we could also try making VLC forwards-compatible with GLSL 130+ and
then enabling the fixed-function ordered dither where available.

Note by Thomas Guillem:
Dithering is disabled by default since it can have an impact on performances
depending on OSses / GL extensions / GPU.

Modified-by: Thomas Guillem <thomas@gllm.fr>
Signed-off-by: Thomas Guillem <thomas@gllm.fr>