macOS codesign.sh: Regenerate the cache after signing modules
The cache is only valid if mod time and size of the plugins match.
As macOS codesign adapts the binaries, the cache will be invalid
after signing. Additionally, the cache file itself must be signed
as well, because of its location inside the bundle. Therefore,
regenerate the cache at the correct stage in the middle of the
signing script.

The script searches for vlc-cache-gen when executed inside vlcs
build directory, as well as inside the release package.

refs #19448