demux:mkv: go one level up in the parser if we found nothing at the current level
It's OK to find nothing if the data was all garbage and dummy we skipped.

The check on the EOF is necessary as there's no way to tell if we found nothing
because of EOF or because we reached the end of the Max size to read without
anything useful.

Fixes #19512

(cherry picked from commit 64f1615ae0627086d5fba94e7bc70f32d494691f)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>