NSIS: Properly detect 32bit VLC
Previously the 64bit installer would not detect if a 32bit VLC
is already installed, this can lead to situations where two VLC
versions are installed on the system, one 32bit and one 64bit.
Windows will only show one version, as they have exactly the same
name in the start menu and shortcuts.
Even though the registry values are different for 32bit and 64bit,
the Apps list of Windows will only show one of them too.
(Except the old non-UWP one)

Therefore we really need to avoid this case. With this commit, if the
user has a 32bit VLC installed, it will be uninstalled first before
installing a 64bit version.