qt: fix recentsMenu that causes skins2 to crash
Skins2 uses FileMenu in the PopupMenu (short lifetime). This causes the
static variable recentsMenu to be freed and then be dereferenced.

The patch creates a one-off recentsMenu, that is used by the different
instances of FileMenu via the addMenu function (no transfer of ownership).

It follows the same pattern as the other two static audioDeviceMenu
and rendererMenu. Note that these static variables are never freed.