chromecast: add a stream out proxy
This new submodule wait for all tracks to be added into the last stream out
(the http one) before running a chromecast session.

This fixes chromecast timeout errors when the session was started during the
middle of a playback and when the http sout was still empty (because no video
IFrames were available for example).

This also fixes the "You can try increasing sout-mux-caching value" error (when
the timeout came from VLC and not the CC).

(cherry picked from commit 5fb95aaf5ea1887e670ee0faf33d71065035f938)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>