input: es_out: compute position from buffering only on update
Position value from es_out depends on buffering but they are
not updated at the same rate (when not byte based). This causes
growing buffering being substracted from a stalled position (based on
pcr, or other interval) and the computed position to regress as the
position isn't updated yet by demuxer.