NSIS: Add custom VLC banner image
This adds a custom VLC banner image to the welcome and finish page of
the NSIS installer and uninstaller.

Note: If there is ever need to re-generate the bmp banner image,
use a custom color pallet, this allows to only need 4 bits per color and
drastically reduces the size. Do not write colorspace information.
Use Run-length coding.