macosx: rework renderer menu to match the Qt interface
Discovery automatically runs for 20s when opening the menu and is terminated afterwards. The state is no longer displayed as it is implicit.

Further, this removes the nib object singleton of the renderer menu controller as this way of initialization is discouraged in modern Cocoa.

(cherry picked from commit f95e46d68d64707906db81c16357192befe775c8)