arm: Allow building assembly in thumb mode
Windows on arm is thumb2 only.

Add the necessary "it" instructions before conditionally executed
instructions (which doesn't emit any extra instructions when not
building in thumb mode). The number of "it" instructions could
be reduced in some places by reordering the instructions, but keeping
them as they were originally to avoid any impact on existing
targets.

Remove redundant .arm directives; the assembler normally starts out
in that mode anyway, and for windows we shouldn't override the
default mode.