macosx: Workaround mkv file association problem
This workarounds a problem very likely within MplayerX. The current
version of that application declares both an exported, as well as
an imported UTI for the mkv file extension, which messes up macOS.
They should delete the exported UTI to fix the problem correctly.

As a workaround, fall back the the old way of declaring mkv file
association, which seems to work as well.

fixes #19758