package/macosx: Update dmg background and icon positions
The previous background had the orange bar at the bottom, but with
some configurations of the Window it would be covered by the status
bar. Therefore move the orange bar to the top and adjust the icon
positions accordingly.

(cherry picked from commit 454623571cee43b12b6377bb76392be35facc0fb)
Signed-off-by: Marvin Scholz <epirat07@gmail.com>