macosx: Save/share time remaining state in fullscreen controller
There was no identifier set for the time remaining label in the
fullscreen controller, this sets one to ensure the state is saved
correctly. Additionally this is the same identifier as the one in
the main window label, so make sure the state is shared between those.

Fix #19813

(cherry picked from commit ecca9675c9dceb3e9ba15f106cac2c26018a5b7f)
Signed-off-by: Marvin Scholz <epirat07@gmail.com>