macosx: Refactor and update drag handling in ConvertAndSave Window
Fixes a crash when dragging a playlist item onto the drag and drop area
of the Convert an Save window, as it tried to access the no longer
present playlist item pointerValue.
Instead, the playlist item id has to be used to get the playlist item.

Additionally refactor the way too complex and hard to follow
draggedItems iteration code.