extras/tools: Patch libtool to find the compiler-rt that clang uses by default
When libtool links DLLs that include C++ code, it does so by
passing -nostdlib and manually adding the compiler built-in
default libraries. When linking, clang specifies the compiler-rt
library by passing the full path to the .a file, instead of using
-L + -l. libtool currently doesn't pick up such a full path to
a static library as a library it should include.

This patch is a custom libtool patch from chromium os [1],
upstreaming is in progress at [2].

[1] https://chromium.googlesource.com/chromiumos/overlays/portage-stable/+/88201a517e3a797e4197bacc4959785a934c08d6
[2] https://debbugs.gnu.org/cgi/bugreport.cgi?bug=27866