mmdevice: handle "--no-volume-save"
It's handled the other way around than old aout modules.

By default, "volume-save" is true and we keep the current behavior (the volume
is saved by Windows for this session). If "volume-save" is false, then we use
the prefered volume set by the user "mmdevice-volume" and don't use the VLC
audio session.

Fixes #19664