config: assert that option exists
We already assert that the type is correct. Only one call site actually
checks for the error return value, and even this, it is debatable
(that is var_Create(VLC_VAR_BOOL)). So an assertion makes more sense
that printing an error.

In some cases (e.g. Lua), we use config_GetType() ahead to check for
errors though - that still works fine.