contrib: ffmpeg: Ensure the securetransport module is disabled
The ffmpeg securetransport module uses private Apple APIs, so it can
cause rejection from the Apple appstore when this symbol is contained
in binaries, even though VLC never uses it in any way directly.
There is no point in enabling this module anyway, a VLC has it's own
implementation of TLS using either GnuTLS or SC, not relying at all on
ffmpeg for this.