auhal: warn about audio configuration one time per device
This warning is quite useful if you don't know how to configure your speakers
to get surround, but it should not be displayed it every time. Indeed, some
users may want to stick to stereo and you can't really differentiate a device
(with more than 2 channels) configured for stereo with a device not configured
at all.

Fixes #19949