Link libcompat with a library with math functions
We detect whether math functions are located in -lm in configure.ac.
Reuse the deliberable of this check and link libcompat with $(LIBM).
This is required at least for sincos() functions and it has been
reported as required on NetBSD 8.0.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>
(cherry picked from commit b6968b7d36769370fe20e095070e665ef2dbf097)
Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>