coreaudio: finer audio delay
This is a partial revert of f18fef325af25113875b09c8f727017febc6198a

This commit fixes the delay imprecision that was depending on the number of
frames asked by the render callback. This also fixes audio glitches (due to
flush/resampling) when this number of frames was changed (when the screen is
OFF for example).