hw: vaapi: fix p010 <=> I420_10 conversions
P010 is on MSB, and VLC_CODEC_I420_10B/VLC_CODEC_I420_10L are on LSB, so we
need to shift bits by 6 when doing the conversion. VLC_CODEC_I420_10L should be
used instead of VLC_CODEC_I420_10B since it's little endian.

It "worked" only because the conversion was tested while downloading and
uploading, so the download error was canceled by the upload error. This fixes
screenshot colors.