contrib: x264: Use a unique file name for the x264 snapshot tarball download
Treat x264 as an (almost) normal package to download, except for not
checking for the download in http://downloads.videolan.org/pub/contrib.

This fixes building contribs with a prepopulated tarballs directory,
even if the x264 snapshot version changes.

Alternatively, the x264 contrib could be changed to use download_git
with proper support for pinning to a specific git hash.

(Not touching the x262 part since that doesn't seem to be working
at all right now. Downloading x262 fails since it tries to download
from an undefined X262_SNAPURL.)