demux: check for pf_readdir directly
Since we anyway test for pf_read and pf_block, and for pf_demux, we
might as well check for pf_readdir being non-NULL directly.

Maybe someone wants to add a VLC stream equivalent for fstat(),
replacing those NULL tests and STREAM_GETSIZE but that is left for
future study.