contrib: ffmpeg: Patch ffmpeg to build with clang without gas-preprocessor
Vanilla clang supports altmacro since clang 5.0, and thus doesn't
require gas-preprocessor for building the arm assembly any longer.

However, the built-in assembler doesn't support a few of the fringe
features used in the assembly.

Patches 1-2 are needed for both libav and ffmpeg, while 3-5 are
needed only on ffmpeg. All patches are already in the latest
corresponding upstream master versions.

Patches 1-2 and 4-5 are only needed on Xcode 9.3, while patch 3
is needed with clang >= 5.0 for any target.