Contribs: activate AOM/AV1
The codec was officially announced, and is almost finished.

This should not be BP to stable until there is an official release of
the bitstream and the software. But it can help to test, for now.

Fixes #21227

(cherry picked from commit 3b0b35f33eeb73b68a037008bdb171a37ae287a1)