Make 1.0x speed accessible from the slider
A slider value change is ignored if the new value is the same as the
last one.

However, the last value was not updated on manual slider move, so its
value remained 0 (speed 2^0 == 1). As a consequence, any slider "move
to 1.0x" event was ignored.

Update the last value on slider move to fix the problem.

Fixes #20170

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>
(cherry picked from commit 6598b2c4d84d6344b413d0cb14563a0a6335956c)
Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>