qt: update all time labels on seek request
There are 3 time labels in the main interface, containing:
 - the elapsed time (on the left of the seekbar);
 - the remaining or total time (on the right of the seekbar);
 - both (in the statusbar).

Only the one from the statusbar was updated on seek request: it was
individually connected to its own specific slot. As a consequence, the
others were not updated on seek (especially visible while paused).

Instead, factorize the slot implementation so that all time label
updates are handled by setDisplayPosition(), and connect seekRequested()
from the TimeLabel constructor (like positionUpdated()) to update them
all on seek request.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>