fingerprinter: fix double-free
In case of thread cancellation (when the audio fingerprinter window is
closed during processing), the current fingerprint request might be left
both in the processing and results queues.

As a consequence, fingerprint_request_Delete() was called twice during
CleanSys(), leading to double-free.

To avoid the problem, remove a request from the processing queue
immediately after it has been added to the results queue, before
vlc_testcancel().

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>