configure: remove check for a header not used in the code
Even if it's not found have_v4l2 will be set to "yes" anyway.

"For each given system header file header-file in the blank-separated argument
list that exists, define HAVE_header-file (in all capitals). If action-if-found
is given, it is additional shell code to execute when one of the header files
is found."