contrib: aom: Update to a recent git snapshot
This version makes sure pthreads-w32 isn't accidentally detected
and enabled; previously when building contribs for windows, aom
would end up trying to use pthreads-w32 (but fail linking the plugin)
if it was built after pthreads-w32.

This version also supports building with assembly or intrinsics
for windows on arm/arm64.

In principle, the same issue also exists in libvpx. However, when
libvpx is built with the contribs build system, the contribs include
directory isn't passed to libvpx, so it won't accidentally find
pthreads-w32. (This issue is fixed in the latest git version of libvpx.)