hotkeys: fix first subtitles toggle
A new hotkey to toggle subtitles visibility was introduced by commit
69d6f2ab18e6956f22923044044b13419a713724. The idea was to save the
current track id every time the subtitles track is switched via the
hotkey (v), and restore it on toggle (Shift+v).

However, the subtitle track is not necessarily changed via the hotkey:
it may be selected as a default subtitle, or switched from the menu for
example. As a consequence, the saved track id may be invalid.

To fix the issue, do not try to save the selected track id every time it
changes; instead, save it on toggle if subtitles are enabled, and
restore the track with the saved id if subtitles are disabled.

Fixes #16721

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit bb21396e96093b30145dfe1ed328853359b1b70c)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>