qt: improve time tooltip for hidpi
The time tooltip (for the seek slider) was pixelated on hidpi screens.

Both the polygon to draw the tooltip and the widget mask had low
definition.

Replace QPolygon by QPolygonF to fix the tooltip drawing.

Currently, Qt does not support hidpi masks. Instead, disable the mask,
set a translucent background and make the widget transparent for mouse
events.

It also fixes #20011 as a side effect.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>