aout: reduce output lock scope
The all audio output calls from the decoder must be (and actually are)
serialized by the decoder.

The output lock needs to protect against concurrent accesses to the
request parameters (owner->req) and the audio output module callbacks,
both of which can be accessed by interface threads.

It does not need to protect anything else, notably synchronization
parameters, the filter chain and the volume mixer.