macosx: Make toolTips on Full Screen consistent
Now the tooltips which are shown in the Normal View are same
as the ones which are shown in Full-Screen View. Adapted the shorter version.
And also, changed the Normal View toolTip of "Fullscreen" to "Enter fullscreen"
fixes #19993

Signed-off-by: David Fuhrmann <dfuhrmann@videolan.org>
(cherry picked from commit c7e23f14795f98bc88475e8eab7c2620212449e9)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>