vdummy: do not delete the window
This means that there will be an unused video window. Insofar as the
dummy video output is meant for debugging/testing, this change can be
considered a bug as much as a feature, depending on the specific
purpose.

The proper way to disable window is still --no-video, which will not
create a video window - no changes there.