window: do not bypass vout thread when resizing
When the vout window is resized, queue an event to the video output
thread. This will wake-up the video output thread automatically and
process the change straight away.

Sending an event directly to the display is somewhat ugly, and will
would not perform any necessary vout thread wake-up.

This removes vout_display_window_Detach(), which had become a no-op.