display: deprecate window state event
In principles, this would tell the core what the window state is,
although that would belong in the window provider.

Currently, this is instead used by a few display plugins to loop an
event back to themselves through the vout thread. The core would be
better left out of this completely.