freetype: Fix compilation on macOS
Usually, the per-target _LINK variable is autogenerated, taking
into account used languages based on file extension, and also
per-target _LDFLAGS. Due to #17827, _LINK for this module is
explicitly overridden, and therefore the per-target _LDFLAGS are not
respected, which leads to compilation failure.

This is a quick fix to make macOS compile again. The problem
affects all platforms however, and a better solution shall be found
later.