display: remove useless continue
Fitting the window will ask for the window to change size (or do
nothing). Typically, the window will not synchronously change size.
But even if it did, it will queue an event to the video output thread
control queue, which only be processed after vout_ManageDisplay()
returns to the thread main loop.

So in any case, by the time the VoutDisplayFitWindow() call returns,
there will be nothing new for vout_ManageDisplay() to do. So there
are no points in looping.

With the continue statement removed, the window fitting block can be
moved out after the loop.