wl_shm: remove position offsets
The correct offsets depend on the direction in which the window was
resized. This cannot be guessed from the change of display size.
(For instance, a reduction in width does not indicate whether it was
 from the left or the right of the window.)

So those offset computations are futile - remove them.
If the offsets are really needed, the window will have to provide the
values (in some form).