contrib: harfbuzz: Skip using the _BitScan intrinsics on mingw
In mingw targets, both when using GCC and Clang, the __GNUC__
intrinsics right above can be used in most cases anyway.

This fixes building with mingw-w64 for armv7/aarch64, which doesn't
implement the _BitScanForward/_BitScanReverse intrinsics yet.

Not upstreaming this fix yet since the correct fix would be to
implement these intrinsics in mingw-w64 for armv7 and aarch64 as well.