xdg-shell: report window size correctly
If a dimension is zero, then we get to choose it, so we pick the last
dimension requested by the core - rather than the current dimension.

If a dimension is non-zero, we have to use the value regardless of what
the core requested. This fixes incorrect video sizes in fullscreen,
maximized or tiled windows.

Note: Some window managers may exhibit a bug here, whereby they do not
reset the window dimensions to zero after leaving fullscreen/maximized
state.