freetype: darwin: Fix heap buffer overflow
The old getCStringCopyForCFStringRef function had a flaw that would
cause a heap-buffer overflow if the CFString easy obtaining of the
pointer fails and the string contains characters that need more than
one byte.

To fix this, remove the old getCStringCopyForCFStringRef and use the
recently added CFStringCopyUTF8CString instead, and generalized it to
work with any encoding. Additionally add a check for the malloc
result to prevent passing NULL to CFStringGetCString.

Fix #20721

(cherry picked from commit e89663939f86bbb4d914060f87f2c129714b2416)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>